changed original Player class to use sqlalchemy Player class.
